+++ src/background/background-manager.test.ts
