+++ src/background/background-manager.ts
