+++ src/background/index.ts
