+++ src/background/tmux-session-manager.test.ts
