+++ src/background/tmux-session-manager.ts
